Implement Strava data sync script and fix activity ID overflow
- Add sync_strava.py: fetch all Strava activities, filter runs, and
  store with auto-generated Chinese description text for future embedding
- Auto-refresh access token on each sync run to avoid 6-hour expiry
- Fix Activity.id column from Integer to BigInteger (Strava IDs exceed
  PostgreSQL int4 max ~2.1 billion)
- Add scripts/__init__.py to enable `python -m scripts.sync_strava`

Tested: successfully synced 822 running records spanning 2021-02 to 2026-03.

+++ app/db/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, DateTime, Text
+from sqlalchemy import Column, BigInteger, Integer, String, Float, DateTime, Text
 from sqlalchemy.orm import declarative_base
 from pgvector.sqlalchemy import Vector
 
@@ -8,7 +8,7 @@ Base = declarative_base()
 class Activity(Base):
     __tablename__ = "activities"
 
-    id = Column(Integer, primary_key=True)  # Strava activity ID
+    id = Column(BigInteger, primary_key=True)  # Strava activity ID
     name = Column(String(255))
     sport_type = Column(String(50))  # Run, TrailRun, etc.
     start_date = Column(DateTime)
